requirements.txt
django-autoslug==1.8.0
django-notifications-hq==0.7.0
djangorestframework==3.1.3
django-extensions==1.5.5
oauthlib==0.7.2
Pillow==2.8.2
psycopg2==2.6.1
pygraphviz==1.2
PyJWT==1.3.0
python-openid==2.2.5
python-social-auth==0.2.11
six==1.9.0
Unipath==1.1
wheel==0.24.0
#sudo apt-get install libjpeg-dev
